:memo: README: Convert to Markdown, add project profile and screenshot (#32)
This commit converts the README to Markdown so it is more beautiful to
read. It also adds a link to the FOSS@RIT project profile introduced
over in FOSSRIT/fossrit.github.io#131. It also embeds a screenshot into
the README, although the screenshot will not render until that PR gets
merged on the website.

Sort of related to FOSSRIT/tasks#113.

Signed-off-by: Justin W. Flory <git@jwf.io>